+++ toolkits/notion/arcade_notion/tools/__init__.py
@@ -1,9 +1,9 @@
-from arcade_notion_toolkit.tools.pages import (
+from arcade_notion.tools.pages import (
     create_page,
     get_page_content_by_id,
     get_page_content_by_title,
 )
-from arcade_notion_toolkit.tools.search import (
+from arcade_notion.tools.search import (
     get_object_metadata,
     get_workspace_structure,
     search_by_title,
