+++ toolkits/notion/arcade_notion/tools/pages.py
@@ -6,12 +6,12 @@ from arcade.sdk import ToolContext, tool
 from arcade.sdk.auth import Notion
 from arcade.sdk.errors import ToolExecutionError
 
-from arcade_notion_toolkit.block_to_markdown_converter import BlockToMarkdownConverter
-from arcade_notion_toolkit.enums import BlockType, ObjectType
-from arcade_notion_toolkit.markdown_to_block_converter import convert_markdown_to_blocks
-from arcade_notion_toolkit.tools.search import get_object_metadata
-from arcade_notion_toolkit.types import DatabaseParent, PageWithPageParentProperties, create_parent
-from arcade_notion_toolkit.utils import (
+from arcade_notion.block_to_markdown_converter import BlockToMarkdownConverter
+from arcade_notion.enums import BlockType, ObjectType
+from arcade_notion.markdown_to_block_converter import convert_markdown_to_blocks
+from arcade_notion.tools.search import get_object_metadata
+from arcade_notion.types import DatabaseParent, PageWithPageParentProperties, create_parent
+from arcade_notion.utils import (
     extract_title,
     get_headers,
     get_next_page,
