+++ toolkits/notion/pyproject.toml
@@ -1,5 +1,5 @@
 [tool.poetry]
-name = "arcade_notion_toolkit"
+name = "arcade_notion"
 version = "0.1.3"
 description = "Arcade.dev LLM tools for Notion"
 authors = ["ArcadeAI <dev@arcade.dev>"]
@@ -21,7 +21,7 @@ requires = ["poetry-core>=1.0.0,<2.0.0"]
 build-backend = "poetry.core.masonry.api"
 
 [tool.mypy]
-files = ["arcade_notion_toolkit/**/*.py"]
+files = ["arcade_notion/**/*.py"]
 python_version = "3.10"
 disallow_untyped_defs = "True"
 disallow_any_unimported = "True"
